Disallow attachments from file:///data/
Disallow third-party applications to send a share intent with a
file:///data/... Uri because that would expose internal application data.

If the Intent Uri is a "file" and the file is in
`Environment.getDataDirectory()`, then throw a AttachmentFailureException
which is caught immediately and shows a toast to the user.

Fix issue 199888
Fix b/26989185

Change-Id: Ice1d3382d52d2ab97cddc3e9c0b2dd29819b664f
(cherry picked from commit 21483a4a65a82563ed4248cef567f547c7446687)